データ追加機能を実装　基本機能実装完了

+++ database/migrations/2022_03_20_022844_create_books_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('books', function (Blueprint $table) {
             $table->unsignedBigInteger('id', true);
-            $table->unsignedInteger('amount');
+            $table->bigInteger('amount');
             $table->boolean('balance'); //0が収入、1が支出
             $table->unsignedBigInteger('time_id');
             $table->unsignedBigInteger('category_id');
